+++ dartagnan/src/main/java/com/dat3m/dartagnan/utils/visualization/ExecutionGraphVisualizer.java
@@ -9,13 +9,20 @@ import com.dat3m.dartagnan.verification.model.ExecutionModel;
 
 import static java.util.Optional.ofNullable;
 
+import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.math.BigInteger;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.TimeUnit;
 import java.util.function.BiPredicate;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 /*
     This is some rudimentary class to create graphs of executions.
@@ -23,6 +30,8 @@ import java.util.function.BiPredicate;
  */
 public class ExecutionGraphVisualizer {
 
+    private static final Logger logger = LogManager.getLogger(ExecutionGraphVisualizer.class);
+	
     private final Graphviz graphviz;
     private BiPredicate<EventData, EventData> rfFilter = (x, y) -> true;
     private BiPredicate<EventData, EventData> coFilter = (x, y) -> true;
@@ -43,16 +52,15 @@ public class ExecutionGraphVisualizer {
     }
 
     public void generateGraphOfExecutionModel(Writer writer, String graphName, ExecutionModel model) throws IOException {
-    	for(Thread t : model.getThreads()) {
-            for(EventData data : model.getThreadEventsMap().get(t)) {
-            	if(data.isMemoryEvent()) {
-            		MemEvent m = (MemEvent)data.getEvent();
-            		if(!(m.getAddress() instanceof Register)) {
-                    	addresses.putIfAbsent(data.getAccessedAddress(), m.getAddress());            			
-            		}
-            	}
-            }
-    	}
+        for(EventData data : model.getThreadEventsMap().values().stream()
+                .collect(ArrayList<EventData>::new, List::addAll, List::addAll)) {
+        	if(data.isMemoryEvent()) {
+        		MemEvent m = (MemEvent)data.getEvent();
+        		if(!(m.getAddress() instanceof Register)) {
+                	addresses.putIfAbsent(data.getAccessedAddress(), m.getAddress());            			
+        		}
+        	}
+        }
         graphviz.begin(graphName);
         graphviz.append(String.format("label=\"%s\" \n", graphName));
         addAllThreadPos(model);
@@ -179,5 +187,27 @@ public class ExecutionGraphVisualizer {
 
     private void appendEdge(EventData a, EventData b, ExecutionModel model, String... options) {
         graphviz.addEdge(eventToNode(a, model), eventToNode(b, model), options);
+    }
+    
+    public static void generateGraphvizFile(ExecutionModel model, int iterationCount, BiPredicate<EventData, EventData> edgeFilter, String directoryName, String fileNameBase) {
+        File fileVio = new File(directoryName + fileNameBase + ".dot");
+        fileVio.getParentFile().mkdirs();
+        try (FileWriter writer = new FileWriter(fileVio)) {
+            // Create .dot file
+            new ExecutionGraphVisualizer()
+                    .setReadFromFilter(edgeFilter)
+                    .setCoherenceFilter(edgeFilter)
+                    .generateGraphOfExecutionModel(writer, "Iteration " + iterationCount, model);
+
+            writer.flush();
+            // Convert .dot file to pdf
+            Process p = new ProcessBuilder()
+                    .directory(new File(directoryName))
+                    .command("dot", "-Tpng", fileNameBase + ".dot", "-o", fileNameBase + ".png")
+                    .start();
+            p.waitFor(1000, TimeUnit.MILLISECONDS);
+        } catch (Exception e) {
+            logger.error(e);
+        }
     }
 }
